refactor: move ArraySum etc to _array_expressions
